+++ position_detection.h
@@ -7,12 +7,12 @@
  * @param position 계산된 x, y 좌표를 저장할 int형 배열 (크기 2)
  * @return 성공 시 1, 실패 시 0을 반환합니다.
  */
-int trilaterate(double distances[4], int position[2]);
+int trilaterate(double distances[4], uint8_t position[2]);
 
 /*
  * @brief UWB 통신을 통해 거리 값을 받아오고, 삼변측량을 통해 현재 위치를 계산하는 전체 과정을 포함하는 함수입니다.
  * @param position 계산된 x, y 좌표를 저장할 포인터 배열
  */
-int update_current_position(int* position);
+int update_current_position(uint8_t* position);
 
 #endif // POSITION_DETECTION_H
